+++ Zilon.Core/Zilon.Emulation.Common/IAutoplayContext.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading.Tasks;
+
+namespace Zilon.Emulation.Common
+{
+    /// <summary>
+    /// The context to handle autoplay state.
+    /// </summary>
+    public interface IAutoplayContext
+    {
+        /// <summary>
+        /// Checks next iteration. If there is no next iteration then autoplay engine stops.
+        /// </summary>
+        Task<bool> CheckNextIterationAsync();
+    }
+}
